requirements.txt
langchain
openai
pinecone-client
pandas
tiktoken
python-dotenv
streamlit